Transformation/requirements.txt
pandas 
SQLalchemy 
psycopg2-binary
python-dotenv 